feat: day 7 and refactor to use i64 return type

+++ contracts/solution/src/lib.rs
@@ -3,6 +3,6 @@ extern crate alloc;
 use alloc::string::String;
 
 pub trait Solution {
-    fn solvepart1(&self, input: String) -> u32;
-    fn solvepart2(&self, input: String) -> u32;
+    fn solvepart1(&self, input: String) -> i64;
+    fn solvepart2(&self, input: String) -> i64;
 }
